Add a reorder module
